Move split_by_chunks function to Util
- Move function to Util module
- Introduce DiscordUtil class in src.backend.discord.utils to keep legacy utils used in Discord specific commands module
- Remove dependency of Util class on Msg class
