+++ src/cmd/reminder.py
@@ -16,7 +16,6 @@ from src.api.command import (
 from src.api.execution_context import ExecutionContext
 from src.api.reminder import Reminder
 from src.backend.discord.embed import DiscordEmbed
-from src.backend.discord.message import Msg
 from src.config import bc
 from src.utils import Time, Util
 
@@ -183,7 +182,7 @@ class _ReminderInternals:
             embed_color = random.randint(0x000000, 0xffffff)
             cur_list = 1
             total_list = int(math.ceil(reminders_count / const.DISCORD_MAX_EMBED_FILEDS_COUNT))
-            for reminder_chunk in Msg.split_by_chunks(reminder_list, const.DISCORD_MAX_EMBED_FILEDS_COUNT):
+            for reminder_chunk in Util.split_by_chunks(reminder_list, const.DISCORD_MAX_EMBED_FILEDS_COUNT):
                 e = DiscordEmbed()
                 if total_list > 1:
                     e.title(f"List of reminders {cur_list}/{total_list}")
